requirements.txt
streamlit>=1.28.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.0.0
pyyaml>=6.0
numpy>=1.21.0
transformers>=4.20.0
torch>=1.12.0
datasets>=2.0.0
scikit-learn>=1.1.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
tqdm>=4.64.0
langdetect>=1.0.9
python-Levenshtein>=0.27.1
openweights